+++ funvas_tweets/lib/src/11.dart
@@ -11,9 +11,7 @@ class Eleven extends Funvas with FunvasTweetMixin {
 
   @override
   void u(double t) {
-    final scaling = min(x.width, x.height) / 750;
-    final w = x.width / scaling, h = x.height / scaling;
-    c.scale(scaling);
+    final s = s2q(750), w = s.width, h = s.height;
 
     c.drawPaint(Paint()..color = const Color(0xfffaddaa));